do not create symlink if itis already there

+++ local/backup/install.sh
@@ -3,4 +3,4 @@
 echo ... installing media backup system
 cp backup-media /usr/local/bin/
 chmod +x /usr/local/bin/backup-media
-ln -s /usr/local/bin/backup-media /usr/local/bin/backup-server
+test -f /usr/local/bin/backup-server || ln -s /usr/local/bin/backup-media /usr/local/bin/backup-server
